Add unit test for all restful controllers

+++ src/main/java/tk/puncha/restfulControllers/RestfulVisitController.java
@@ -9,6 +9,7 @@ import tk.puncha.models.Visit;
 import tk.puncha.repositories.PetRepository;
 import tk.puncha.repositories.VisitRepository;
 
+import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 
 @RestController  // Includes @ResponseBody
@@ -36,13 +37,20 @@ public class RestfulVisitController {
   }
 
   @PostMapping
+  @ResponseStatus(HttpStatus.CREATED)
   public void create(Pet pet, @Valid @RequestBody Visit visit) {
+    if(pet == null)
+      throw new EntityNotFoundException("Pet doesn't exist.");
     visit.setPet(pet);
     visitRepository.insert(visit);
   }
 
   @DeleteMapping("{id}")
-  public void delete(@PathVariable int id) {
+  @ResponseStatus(HttpStatus.NO_CONTENT)
+  public void delete(Pet pet, @PathVariable int id) {
+    if(pet == null)
+      throw new EntityNotFoundException("Pet doesn't exist.");
+
     visitRepository.deleteById(id);
   }
 }
